api-scd/requirements.txt
fastapi 
uvicorn 
Pillow
numpy
pydantic
fastapi[standard]